+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.4.1
+- Fixed a compatibility issue with libWrapper.
+- Fixed resource creation when there are less than two.
+
+1.4.0
+- Updated for FoundryVTT 0.9.
+- Added separate visibility configuration for owner and everyone else.
+- Added option to store default settings per actor type.
+- Fixed defaults not applying to prototype tokens.
+- It is now possible to load empty defaults using the button. Empty defaults will not be applied when creating actors and tokens.
+
 1.3.1
 - Fixed a bug that caused bars to appear twice.
 - Fixed advanced bar configuration for prototype tokens.